+++ src/DisAsm/Reader.h
@@ -18,8 +18,8 @@ typedef void * HREADER;
 struct ReaderContext_t;
 
 typedef int  (*pfnRead)(struct ReaderContext_t * hReader, void * buffer, uint32_t size);
-typedef int  (*pfnSeek)(struct ReaderContext_t * hReader, uint32_t pos);
-typedef int  (*pfnSkip)(struct ReaderContext_t * hReader, uint32_t count);
+typedef int  (*pfnSeek)(struct ReaderContext_t * hReader, uint64_t pos);
+typedef int  (*pfnSkip)(struct ReaderContext_t * hReader, uint64_t count);
 typedef void (*pfnDestroy)(struct ReaderContext_t * hReader);
 
 typedef struct ReaderContext_t
@@ -33,8 +33,8 @@ typedef struct ReaderContext_t
 ReaderContext;
 
 int ReaderRead(HREADER hReader, void * buffer, uint32_t size);
-int ReaderSeek(HREADER hReader, uint32_t pos);
-int ReaderSkip(HREADER hReader, uint32_t count);
+int ReaderSeek(HREADER hReader, uint64_t pos);
+int ReaderSkip(HREADER hReader, uint64_t count);
 void ReaderDestroy(HREADER hReader);
 
 #endif /* __READER_H__ */
